requirements.txt
# Google AI Development Kit (ADK) dependencies
# Conditional installation based on Python version
google-adk==0.3.0; python_version<"3.9"
google-adk==1.15.1; python_version>="3.9"

# Additional dependencies that might be needed
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest==8.4.2
pytest-asyncio==1.2.0
black>=23.0.0
flake8>=6.0.0
